+++ client-sdk/ts-web/core/src/client.ts
@@ -1,5 +1,6 @@
 import * as grpcWeb from 'grpc-web';
 
+import * as proto from '../proto';
 import * as misc from './misc';
 import * as types from './types';
 
@@ -538,6 +539,15 @@ const methodDescriptorDebugControllerWaitNodesRegistered = createMethodDescripto
     void
 >('DebugController', 'WaitNodesRegistered');
 
+// see oasis-core/go/common/grpc/errors.go
+/**
+ * grpcError is a serializable error.
+ */
+interface GRPCError {
+    module?: string;
+    code?: number;
+}
+
 export class GRPCWrapper {
     client: grpcWeb.AbstractClientBase;
     base: string;
@@ -560,6 +570,22 @@ export class GRPCWrapper {
                 // todo: unhack this when they release with our change
                 // https://github.com/grpc/grpc-web/pull/1025
                 return undefined;
+            }
+            if (e.metadata && 'grpc-status-details-bin' in e.metadata) {
+                const statusU8 = misc.fromBase64(e.metadata['grpc-status-details-bin']);
+                const status = proto.google.rpc.Status.decode(statusU8);
+                const details = status.details;
+                // `errorFromGrpc` from oasis-core checks for exactly one entry in Details.
+                // We additionally check that the type URL is empty, consistent with how
+                // `errorToGrpc` leaves it blank.
+                if (details.length === 1 && details[0].type_url === '' && details[0].value) {
+                    const grpcError = misc.fromCBOR(details[0].value) as GRPCError;
+                    if (!e.message) {
+                        e.message = `Message missing, module=${grpcError.module} code=${grpcError.code}`;
+                    }
+                    e.oasisCode = grpcError.code;
+                    e.oasisModule = grpcError.module;
+                }
             }
             // grpc-web gives us a plain object with no stack trace, which is less helpful.
             // Transfer it to an Error.