+++ client-sdk/ts-web/core/docs/changelog.md
@@ -10,6 +10,7 @@ New features:
 
 - `oasis.misc` now has `toBase64`, `fromBase64`, and `toStringUTF8`.
 - Errors from GRPC are now wrapped to include a stack trace.
+- Errors from oasis-core with code and module now have `oasisCode` and `oasisModule` fields.
 
 Bug fixes: